Make Repository*Commands subclasses of the RepositoryCommand
